+++ public/legal.php
@@ -1,0 +1,40 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Legal notice</title>
+    <!-- STYLES -->
+    <link rel="preconnect" href="https://cdn.jsdelivr.net">
+    <link rel="dns-prefetch" href="https://cdn.jsdelivr.net">
+    <link rel="preload" href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
+</head>
+<body>
+    <div class="container-md">        
+        <!-- GO BACK LINK -->
+        <a href="./form.php">Return to the form</a>
+
+        <!-- TITLE -->
+        <h1 class="h1 text-center">Legal notice</h1>
+        <h2 class="h4 text-center">This is a sort of warning, and a heads up at some possible problems along the way.</h2>
+
+        <!-- SECURITY -->
+        <h3 class="h3">Security</h3>
+        <p>
+            I am, in no way, responsible for the securiy of the system, or the damage it may cause, wether directly or indirectly. Nor am I responsible for any sort of exploiting it may cause on other infraestructures.
+        </p>
+
+        <!-- ILLEGAL PRACTICES -->
+        <h3 class="h3">Illegal practices</h3>
+        <p>
+            You can get some very basic information, that information may be private, or not to be fetched through public means. Licenses still hold their own, if you screenshot, for example, Google Images, those images have licenses, you're not buying that license, and you will be responsible for a bad use, not me.
+        </p>
+        <h4 class="h4">Advice</h4>
+        <p>
+            This is a PoC, use at your own risky, highly recomended to use for your own resources from which you hold all the rights.
+        </p>
+    </div>
+</body>
+</html>